I finished main function

+++ src/main/java/com/zhytelnyi/coursework/queueapp/controllers/QueueController.java
@@ -63,25 +63,47 @@ public class QueueController extends HttpServlet {
             case "deleteUserByName":
                 deleteUserByName(request, response);
                 break;
+            case "writeControl":
+                writeControl(request, response);
+                break;
             default:
                 response.sendError(400, "wrong command");
         }
     }
 
-    private void deleteUserByName(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    private void writeControl(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        if (!authenticateAndAuthorize(request, response)) {
+            return;
+        }
+        String command = request.getParameter("value");
         Long id = Long.valueOf(request.getParameter("id"));
-        String name = request.getParameter("name");
-        queueDao.deleteUserByName(id, name);
+        queueDao.setAccessQueue(id, command);
+
+
+
         response.sendRedirect("index.jsp");
 
+    }
+
+    private void deleteUserByName(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        if (!authenticateAndAuthorize(request, response)) {
+            return;
+        }
+        String name = request.getParameter("name");
+        Long id = Long.valueOf(request.getParameter("id"));
+        queueDao.deleteUserByName(id, name);
+        response.sendRedirect("index.jsp");
     }
 
     private void nextUser(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        if (!authenticateAndAuthorize(request, response)) {
+            return;
+        }
         Long id = Long.valueOf(request.getParameter("id"));
         queueDao.nextUser(id);
         response.sendRedirect("index.jsp");
-
     }
+
 
     private void manage(HttpServletRequest request, HttpServletResponse response) throws IOException {
         Long id = Long.valueOf(request.getParameter("id"));
@@ -94,10 +116,15 @@ public class QueueController extends HttpServlet {
     }
 
     private void addUser(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        Long id = Long.valueOf(request.getParameter("id"));
+
+        if (!queueDao.isQueueAvailableForWrite(id)) {
+            response.sendError(HttpServletResponse.SC_FORBIDDEN);
+            return;
+        }
 
         String userName = request.getParameter("userName");
         System.out.println(userName);
-        Long id = Long.valueOf(request.getParameter("id"));
 
         if (userName == null) {
             try {
@@ -135,7 +162,8 @@ public class QueueController extends HttpServlet {
                 throw new RuntimeException(e);
             }
         } else {
-            queueDao.create(new QueueModel(name));
+            String username = request.getUserPrincipal().getName();
+            queueDao.create(new QueueModel(name, username));
             response.sendRedirect("index.jsp");
         }
     }
@@ -145,17 +173,9 @@ public class QueueController extends HttpServlet {
 
     private void delete(HttpServletRequest request, HttpServletResponse response) throws IOException {
         try {
-            // Check if the user is authenticated
-            if (!request.authenticate(response)) {
-                // If the user is not authenticated, stop processing the request
+            if (!authenticateAndAuthorize(request, response)) {
                 return;
             }
-            if (!request.isUserInRole("queue_owner")) {
-                // If the user does not have the "queue_owner" role, send a 403 Forbidden error
-                response.sendError(HttpServletResponse.SC_FORBIDDEN);
-                return;
-            }
-            // If the user is authenticated, continue processing the request
             Long id = Long.valueOf(request.getParameter("id"));
             queueDao.deleteById(id);
             response.sendRedirect("index.jsp");
@@ -169,6 +189,30 @@ public class QueueController extends HttpServlet {
         User user = session.getUser();
 //        user.createQueue();
     }
+
+    private boolean authenticateAndAuthorize(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        try {
+            if (!request.authenticate(response)) {
+                return false;
+            }
+        } catch (ServletException e) {
+            throw new RuntimeException(e);
+        }
+        if (!request.isUserInRole("queue_owner")) {
+            response.sendError(HttpServletResponse.SC_FORBIDDEN);
+            return false;
+        }
+        String username = request.getUserPrincipal().getName();
+        Long id = Long.valueOf(request.getParameter("id"));
+
+        Boolean check = queueDao.checkProperty(id, username);
+        if (!check) {
+            response.sendError(HttpServletResponse.SC_FORBIDDEN);
+            return false;
+        }
+        return true;
+    }
+
 
 
 }
